pyproject.toml
[tool.poetry]
name = "healthy-habits-tracker"
version = "0.1.0"
description = ""
authors = ["angelina <angln.pnn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.10"
psycopg2 = "^2.9.9"
pillow = "^10.2.0"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.1"
redis = "^5.0.2"
django-filter = "^23.5"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.4.3"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.1"
requests = "^2.31.0"
stripe = "^8.5.0"
celery = "^5.3.6"
flake8 = "^7.0.0"
django-celery-beat = "^2.5.0"
load-dotenv = "^0.1.0"
aiogram = "2.14.3"
async = "^0.6.2"
executor = "^23.2"
asyncio = "^3.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
